// out/model/operationForPost.ts
/**
 * TransferTo Calculator API
 * TransferTo Calculator
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface OperationForPost {
    /**
     * The operator of this operation
     */
    operator?: string;
    /**
     * The second number of the operation
     */
    number1?: number;
    /**
     * The first number of the operation
     */
    number0?: number;
}